+++ src/Mapping/mapping_routines.h
@@ -8,7 +8,7 @@ libmapping_project_data_to_matrices
     int nsamples , int ndets, 
     double phi[], double theta[], double psi[],  
     double det_pol_angles[],
-    float data[] , int bad_data_samples[], 
+    double data[] , int bad_data_samples[], 
     int dets_to_map[], 
     int map_nside,                                                                                            
     int map_size , int pixels_in_the_map[],
